Transposition table for minimax AI

+++ include/ai/search_position.hpp
@@ -22,7 +22,7 @@ public:
     /**
      * @return Position evaluation from the perspective of side to move.
      */
-    double get_eval() const;
+    int32_t get_eval() const;
 
     /**
      * @param ordered whether to apply move ordering
@@ -54,7 +54,7 @@ private:
      * @param type piece type
      * @return Evaluation material value for the piece.
      */
-    double _material_value(PieceType type) const;
+    int32_t _material_value(PieceType type) const;
 
     /**
      * Piece-Square Table value.
@@ -63,17 +63,17 @@ private:
      * @param square square index 0-63 (8 * rank + file)
      * @return Evaluation value for the piece on the square.
      */
-    double _pst_value(PieceType type, PlayerColor color, int square) const;
+    int32_t _pst_value(PieceType type, PlayerColor color, int square) const;
 
     /**
      * Calculate the full evaluation from the current position.
      * @return The evaluation of the board (from white's perspective).
      */
-    double _compute_full_eval();
+    int32_t _compute_full_eval();
 
 private:
     Board m_board;
 
-    double m_eval; // Evaluation from white's perspective
-    std::vector<double> m_eval_history;
+    int32_t m_eval; // Evaluation from white's perspective
+    std::vector<int32_t> m_eval_history;
 };
